Add code and tables for Venn diagrams. Refactor GO and table reading usage
Finish refactoring code to separate location, begin Melanie tasks

Add up/down info to venn diagram genes

Add extra columns to venn diagram tables, refactor readers
