+++ src/module_overlap/module_deg_overlap.py
@@ -4,9 +4,7 @@
 Filter DEG representation data. Determine overlap with modules from
 WGCNA.
 
-Create a line plot of DEGs over the time points
-Create a Venn Diagram of DEGs over the time points to see overlap between
-Liberty and Draper datasets
+What modules have the highest or lowest amount of DEGs?
 """
 
 __author__ = "Scott Teresi"
@@ -17,138 +15,19 @@ import logging
 import numpy as np
 import pandas as pd
 import os
-import matplotlib.pyplot as plt
-from matplotlib_venn import venn2, venn2_circles, venn2_unweighted
 from functools import reduce
 
-from src.modules.filter_modules import read_gene_modules_table
 from src.module_overlap.module_log2fc_overlap import (
     calc_gene_counts_per_module,
     top_n_modules_and_percent_per_context,
 )
 
-
-def read_DEG_dir(input_directory):
-    """
-    Reads an input directory and returns a list of file paths that contain
-    the magic substring 'Direction.tsv'.
-
-    Args:
-        input_directory (str): Path to directory that contains files with our
-        magic substring. This directory is parsed to find filenames that match
-        the substring.
-
-    Returns:
-        deg_files (list): List of absolute file paths
-    """
-    deg_files = [
-        os.path.join(input_directory, f)
-        for f in os.listdir(input_directory)
-        if (os.path.isfile(os.path.join(input_directory, f))) and ("Direction" in f)
-    ]  # MAGIC substring for filename recognition
-
-    if len(deg_files) != 28:  # MAGIC number of files, anymore or less will
-        # result in error
-        raise ValueError(
-            """More files were tagged than expected, check to see
-                         if files are being tagged appropriately."""
-        )
-
-    return deg_files
-
-
-def read_all_DEG_direction_files(deg_files):
-    """
-    Reads a list of file paths and reads them as pandas files. Creates a list
-    of pandas dataframes. Magic column names are read from the file
-
-    Args:
-        deg_files (str): Output from read_log_2fc_dir function. List
-        of file paths.
-
-    Returns:
-        list_of_read_deg_files (list): List of pandas dataframes
-    """
-    list_of_read_deg_files = []
-    for deg_path in deg_files:
-        individual_deg_file = pd.read_csv(deg_path, sep="\t", header="infer")
-
-        if len(os.path.splitext(os.path.basename(deg_path))) > 2:
-            # MAGIC remove filename extension, does not work if multiple periods
-            # are in filename.
-            raise ValueError(
-                """There were multiple periods in your file path,
-                             this will cause an error."""
-            )
-        analysis_context = os.path.splitext(os.path.basename(deg_path))[0]
-        analysis_context = analysis_context.rstrip("_Direction")  # NB strip
-        # the string from the end so the groupby unstack command in
-        # convert_direction_integer_to_string gives us a more appropriate
-        # column name
-
-        individual_deg_file.rename(
-            columns={
-                "Direction_Differentially_Regulated": analysis_context,
-                "Gene_Name": "Blueberry_Gene",
-            },
-            inplace=True,
-        )  # MAGIC column name
-        individual_deg_file = convert_direction_integer_to_string(
-            individual_deg_file, analysis_context
-        )
-        list_of_read_deg_files.append(individual_deg_file)
-
-    return list_of_read_deg_files
-
-
-def panda_info(panda_file):
-    print(panda_file)
-    print(panda_file.shape)
-    print(panda_file.columns)
-    print(panda_file.info())
-
-
-def convert_direction_integer_to_string(deg_panda_file, analysis_context):
-    """
-    Take the output from EdgeR (the direction files), where a gene can either
-    be 0, 1, or -1 and convert those integer files to strings signifying the
-    direction of regulation (i.e up, down, or no change)
-
-    Args:
-        deg_panda_file (pandas.DataFrame):
-            Index:
-                RangeIndex
-            Columns:
-                Blueberry_Gene: Object, dtype: object
-                analysis_context: Integer, dtype: int64
-            Shape:
-                (93583, 2)
-
-        analysis_context (string):
-            String representing the column name and comparison context
-            of the EdgeR data.
-
-    Returns: deg_panda_file (pandas.DataFrame):
-        Index:
-            RangeIndex
-        Columns:
-            Blueberry_Gene: Object, dtype: object
-            analysis_context + _'Direction': Object, dtype: object
-        Shape:
-            (93583, 2)
-    """
-    deg_panda_file.loc[
-        deg_panda_file[analysis_context] == -1, analysis_context + "_Direction"
-    ] = "Down"
-    deg_panda_file.loc[
-        deg_panda_file[analysis_context] == 1, analysis_context + "_Direction"
-    ] = "Up"
-    deg_panda_file.loc[
-        deg_panda_file[analysis_context] == 0, analysis_context + "_Direction"
-    ] = "No_Change"
-    deg_panda_file.drop(columns=analysis_context, inplace=True)
-
-    return deg_panda_file
+from src.read_tables_and_dir import (
+    read_DEG_dir,
+    read_all_DEG_direction_files,
+    read_synteny_homology_table,
+    read_gene_modules_table,
+)
 
 
 def count_direction_of_regulation_per_module_deg(all_merged):
@@ -208,216 +87,6 @@ def count_direction_of_regulation_per_module_deg(all_merged):
     return counts
 
 
-def get_unique_and_non_unique_genes_per_comparison(
-    all_degs_pandas, draper_col, liberty_col
-):
-    """
-    Gets the counts of unique and non-unique DEGs per genotype comparison.
-    Basically does Draper C1 and T1 vs Liberty C1 and T1. I am grouping the
-    DEGs of up and down for each genotype to get the overall. Then I use set
-    terminology to determine which genes are shared and unshared for each
-    comparison.
-
-    Args:
-        all_degs_pandas (pandas.Data.Frame): Pandas dataframe of
-        blueberry genes and DEG comparison columns. Values are Down, Up, or
-        No_Change to signify what kind of DEG a gene is.
-            Columns:
-                Blueberry_Gene: Object, dtype: object
-                Expression Contexts (MAGIC VARIES): Object: dtype: object
-
-        draper_col (str): String of column name to subset out of the
-        all_degs_pandas pandas dataframe
-
-        liberty_col (str): String of column name to subset out of the
-        all_degs_pandas pandas dataframe
-
-    Returns:
-        awful tuple lol please refactor if using in the future:
-            count_of_common_genes (int):
-            count_of_unique_draper_genes (int)
-            count_of_unique_liberty_genes (int)
-    """
-    # Get smaller panda dataframes with only the relevant data
-    all_changes_draper_series = all_degs_pandas[draper_col].loc[
-        all_degs_pandas[draper_col] != "No_Change"
-    ]
-    all_changes_liberty_series = all_degs_pandas[liberty_col].loc[
-        all_degs_pandas[liberty_col] != "No_Change"
-    ]
-    common_genes = set(all_changes_draper_series.index).intersection(
-        set(all_changes_liberty_series.index)
-    )
-    unique_liberty_genes = set(all_changes_liberty_series.index) - common_genes
-    unique_draper_genes = set(all_changes_draper_series.index) - common_genes
-
-    return (
-        common_genes,
-        unique_draper_genes,
-        unique_liberty_genes,
-    )
-
-
-def graph_venn_diagram_deg_per_context(all_degs_pandas, output_dir):
-    """
-    Create a Venn diagram of the TOTAL DEGs for each genotype at each stage.
-
-    Args:
-        all_degs_pandas (pandas.Data.Frame): Pandas dataframe of
-        blueberry genes and DEG comparison columns. Values are Down, Up, or
-        No_Change to signify what kind of DEG a gene is.
-            Columns:
-                Blueberry_Gene: Object, dtype: object
-                Expression Contexts (MAGIC VARIES): Object: dtype: object
-
-        output_dir (str): Path
-
-    Returns: None, just makes graphs and saves them to the output dir.
-    """
-    all_columns = all_degs_pandas.columns.to_list()
-    all_columns.remove("Blueberry_Gene")
-    all_degs_pandas.set_index("Blueberry_Gene", inplace=True)
-    columns = sorted(all_degs_pandas.columns.to_list())
-    only_draper_colnames = [genome for genome in columns if "LIB" not in genome]
-    only_liberty_colnames = [genome for genome in columns if "DRA" not in genome]
-
-    for draper_col, liberty_col in zip(only_draper_colnames, only_liberty_colnames):
-        (
-            common_genes,
-            unique_draper_genes,
-            unique_liberty_genes,
-        ) = get_unique_and_non_unique_genes_per_comparison(
-            all_degs_pandas, draper_col, liberty_col
-        )
-        count_of_common_genes = len(common_genes)
-        count_of_unique_draper_genes = len(unique_draper_genes)
-        count_of_unique_liberty_genes = len(unique_liberty_genes)
-
-        day_number = int(draper_col.rstrip("_Direction")[-1])  # MAGIC
-        # Magic number just to get the Day integer
-        comparison_name = (
-            "Day %s: All DEGS of Resistant Genotype (Draper) vs. All DEGs of Susceptible Genotype (Liberty)"
-            % day_number
-        )
-
-        comparison_filename = "of_Day_" + str(day_number) + ".tsv"
-        # NOTE THIS is ugly.
-        with open(
-            os.path.join(output_dir, "Shared_DEGS_" + comparison_filename), "w"
-        ) as f:
-            for line in common_genes:
-                f.write("%s\n" % line)
-        with open(
-            os.path.join(output_dir, "Unique_Draper_DEGS_" + comparison_filename), "w"
-        ) as f:
-            for line in unique_draper_genes:
-                f.write("%s\n" % line)
-        with open(
-            os.path.join(output_dir, "Unique_Liberty_DEGS_" + comparison_filename), "w"
-        ) as f:
-            for line in unique_liberty_genes:
-                f.write("%s\n" % line)
-
-        the_labels = ("Unique Draper DEGs", "Unique Liberty DEGs")
-        venn2(
-            subsets=(
-                count_of_unique_draper_genes,
-                count_of_unique_liberty_genes,
-                count_of_common_genes,
-            ),
-            set_colors=("b", "r"),
-            set_labels=the_labels,
-        )
-
-        # Add another set to make the edges bold. Purely cosmetic
-        venn2_circles(
-            subsets=(
-                count_of_unique_draper_genes,
-                count_of_unique_liberty_genes,
-                count_of_common_genes,
-            )
-        )
-        plt.title(comparison_name)
-        plt.savefig(
-            os.path.join(output_dir, "Day" + str(day_number) + "_DEG_VennDiagram.png"),
-            bbox_inches="tight",
-        )
-        plt.clf()
-    return None
-
-
-def graph_line_plot_deg_per_context(merged_deg_no_module_ID, output_dir):
-    """
-    Args:
-        merged_deg_no_module_ID (pandas.Data.Frame): Pandas dataframe of
-        blueberry genes and DEG comparison columns. Values are Down, Up, or
-        No_Change to signify what kind of DEG a gene is.
-            Columns:
-                Blueberry_Gene: Object, dtype: object
-                Expression Contexts (MAGIC VARIES): Object: dtype: object
-
-        output_dir (str): Path
-    """
-    all_columns = merged_deg_no_module_ID.columns.to_list()
-    all_columns.remove("Blueberry_Gene")
-    merged_deg_no_module_ID.set_index("Blueberry_Gene", inplace=True)
-    data = merged_deg_no_module_ID.apply(pd.value_counts)
-    data.fillna(0, inplace=True)
-    columns = sorted(data.columns.to_list())
-    only_draper = [genome for genome in columns if "LIB" not in genome]
-    only_liberty = [genome for genome in columns if "DRA" not in genome]
-
-    # Draper
-    draper_count_dataframe = data[only_draper]
-    down_sequence_draper = draper_count_dataframe.loc["Down"].to_list()
-    up_sequence_draper = draper_count_dataframe.loc["Up"].to_list()
-
-    # Liberty
-    liberty_count_dataframe = data[only_liberty]
-    down_sequence_liberty = liberty_count_dataframe.loc["Down"].to_list()
-    up_sequence_liberty = liberty_count_dataframe.loc["Up"].to_list()
-
-    x_range = ["C" + str(i) + " vs. " + "T" + str(i) for i in range(1, 8, 1)]
-    M = 10
-    plt.figure(figsize=(16, 9.5))
-    plt.plot(
-        x_range,
-        up_sequence_draper,
-        "^b:",
-        label="Up-Regulated in Resistant Genotype (Draper)",
-        markersize=M,
-    )
-    plt.plot(
-        x_range,
-        down_sequence_draper,
-        "vb-",
-        label="Down-Regulated in Resistant Genotype (Draper)",
-        markersize=M,
-    )
-    plt.plot(
-        x_range,
-        up_sequence_liberty,
-        "^r:",
-        label="Up-Regulated in Susceptible Genotype (Liberty)",
-        markersize=M,
-    )
-    plt.plot(
-        x_range,
-        down_sequence_liberty,
-        "vr-",
-        label="Down-Regulated in Susceptible Genotype (Liberty)",
-        markersize=M,
-    )
-    plt.ylabel("Number of Genes")
-    plt.xlabel("Comparison Context")
-    plt.title("Number of Genes Differentially Expressed Per Context")
-    plt.legend()
-    plt.savefig(
-        os.path.join(output_dir, "Lineplot_number_diffex_genes_per_context.png"),
-        bbox_inches="tight",
-    )
-
-
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="")
     parser.add_argument("DEG_dir", type=str, help="output from EdgeR")
@@ -426,12 +95,19 @@ if __name__ == "__main__":
         type=str,
         help="file representing the blueberry genes and their module IDs",
     )
+
+    parser.add_argument(
+        "ortholog_pairs",
+        type=str,
+        help="path to synteny/homology file containing AT BB pairs",
+    )
     parser.add_argument("output_dir", type=str, help="output_dir path")
 
     # NB set args
     args = parser.parse_args()
     args.DEG_dir = os.path.abspath(args.DEG_dir)
     args.genes_and_module_colors = os.path.abspath(args.genes_and_module_colors)
+    args.ortholog_pairs = os.path.abspath(args.ortholog_pairs)
     args.output_dir = os.path.abspath(args.output_dir)
 
     # NB set logging
@@ -446,6 +122,7 @@ if __name__ == "__main__":
     gene_counts_per_module = calc_gene_counts_per_module(
         blueberry_genes_and_module_colors_table
     )
+    synteny_homology_table = read_synteny_homology_table(args.ortholog_pairs)
 
     # NOTE begin the analysis
     list_of_deg_paths = read_DEG_dir(args.DEG_dir)
@@ -454,15 +131,6 @@ if __name__ == "__main__":
     merged_deg_no_module_ID = reduce(
         lambda left, right: pd.merge(left, right, on="Blueberry_Gene", how="outer"),
         list_of_deg_pandas,
-    )
-
-    # NOTE this is a unique special result folder
-    graph_venn_diagram_deg_per_context(
-        merged_deg_no_module_ID.copy(deep=True),
-        os.path.join(args.output_dir, "Unique_and_Shared_DEGs"),
-    )
-    graph_line_plot_deg_per_context(
-        merged_deg_no_module_ID.copy(deep=True), args.output_dir
     )
 
     merged_deg_w_module_ID = pd.merge(
@@ -478,14 +146,18 @@ if __name__ == "__main__":
     modules_and_count_per_context = count_direction_of_regulation_per_module_deg(
         degs_and_modules,
     )
+
+    # Fill the cells with NaN as 0 since we didn't get any genes for that given
+    # module and expression context
+    modules_and_count_per_context.fillna(0, inplace=True)
+
     # Drop the 'No_Change' columns
     list_of_no_change_cols = [
         f for f in modules_and_count_per_context.columns if "No_Change" in f
     ]
     modules_and_count_per_context.drop(columns=list_of_no_change_cols, inplace=True)
 
-    # Get the gene counts per module information again because it was lost in
-    # the previous function.
+    # Add counts per module to the table
     modules_and_count_per_context = modules_and_count_per_context.merge(
         gene_counts_per_module, on="Module_Color", how="outer"
     )
@@ -499,7 +171,6 @@ if __name__ == "__main__":
         columns=["Number_of_Blueberry_Genes_in_Module"], inplace=True
     )  # NB remove column
     percentages = percentages.add_prefix("Percent_")  # NB add prefix to column
-
     # NOTE FUTURE
     # Get gene counts per module again because the number of blueberry genes
     # per module column was set to 1 because of division, in future figure out
@@ -509,7 +180,7 @@ if __name__ == "__main__":
     percentages = percentages.merge(
         gene_counts_per_module, on="Module_Color", how="outer"
     )
-    percentages.fillna(np.nan, inplace=True)
+
     percentages.set_index("Module_Color", inplace=True)
     top_n_modules_and_percent_per_context(percentages, args.output_dir)
